requirements.txt
# Vietnam Real Estate Web Crawler - Requirements

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Scheduling
APScheduler==3.10.4
pytz==2023.3

# Data processing
pandas==2.1.3
pyarrow==13.0.0

# Database
SQLAlchemy==2.0.23

# Environment
python-dotenv==1.0.0

# Logging and monitoring
python-logging-loki==0.3.2

# Utilities
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-timeout==2.2.0

# Code quality
flake8==6.1.0
black==23.12.0
pylint==3.0.3

# Type checking
mypy==1.7.1

# Development
ipython==8.18.1
ipdb==0.13.13
